requirements.txt
fastapi
python-jose[cryptography]
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx
pytest
python-multipart